requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
groq>=0.5.0
python-multipart>=0.0.9
pydantic>=2.6.0
email-validator>=2.1.1
pdfplumber>=0.10.0
python-docx>=0.8.11
argon2-cffi>=21.3.0
